+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 
 * Added additional documentation to function **joes_giant_toolbox.google_cloud.move_or_rename_file_in_gcloud_bucket**
 
+* Added class **joes_giant_toolbox.dataviz.PythonPlottingTutorials**
+
+* Added documentation to **joes_giant_toolbox.dataviz**
+
+* Added documentation to **joes_giant_toolbox.google_cloud**
+
 * Added package dependency group [google]
 
 * Improved (verbose=True) output of function **joes_giant_toolbox.google_cloud.move_or_rename_file_in_gcloud_bucket**